git-svn-id: https://svn.astron.nl/LOFAR/trunk/CEP/PyBDSM@36668 b3281354-443f-0410-9bd6-bf5212d75c11

+++ src/python/rmsimage.py
@@ -951,7 +951,7 @@ class Op_rmsimage(Op):
             i1 = arr[i]
             i2 = arr[i+1]
             t = N.arange(ceil(i1), floor(i2)+1, dtype=float)
-            res[ceil(i1):floor(i2)+1] = i + (t-i1)/(i2-i1)
+            res[int(ceil(i1)):int(floor(i2))+1] = i + (t-i1)/(i2-i1)
 
         return res
